Added CSA file reference. Cleaned up CSA goodie bag. Updated README.
Added a list detailing the files a CSA might need during an event. Have
the files as well, but looking for a home for them as GitHub has a soft
limit of 1GB.

Moved flash drive info to top of CSA goodie bag list in order to clean
up the over layout.

Updated README with details on the changes to the repo, eg; csa files.